Add tests for Line style page
